# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_Zapbox-API_evolution-go_pkg_message_service.ChatPresenceStruct:
    properties:
      isAudio:
        type: boolean
      number:
        type: string
      state:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_message_service.DownloadImageStruct:
    properties:
      directPath:
        type: string
      fileEncSHA256:
        items:
          type: integer
        type: array
      fileLength:
        type: integer
      fileSHA256:
        items:
          type: integer
        type: array
      mediaKey:
        items:
          type: integer
        type: array
      mimetype:
        type: string
      url:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_message_service.EditMessageStruct:
    properties:
      chat:
        type: string
      message:
        type: string
      messageId:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_message_service.MarkReadStruct:
    properties:
      id:
        items:
          type: string
        type: array
      number:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_message_service.MessageStatusStruct:
    properties:
      id:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_message_service.MessageStruct:
    properties:
      chat:
        type: string
      messageId:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_message_service.ReactStruct:
    properties:
      id:
        type: string
      number:
        type: string
      reaction:
        type: string
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
...
  /message/delete:
    post:
      consumes:
      - application/json
      description: Delete a message for everyone
      parameters:
      - description: Delete a message for everyone
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_message_service.MessageStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a message for everyone
      tags:
      - Message
  /message/downloadimage:
    post:
      consumes:
      - application/json
      description: Download an image
      parameters:
      - description: Download an image
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_message_service.DownloadImageStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Download an image
      tags:
      - Message
  /message/edit:
    post:
      consumes:
      - application/json
      description: Edit a message
      parameters:
      - description: Edit a message
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_message_service.EditMessageStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Edit a message
      tags:
      - Message
  /message/markread:
    post:
      consumes:
      - application/json
      description: Mark a message as read
      parameters:
      - description: Mark a message as read
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_message_service.MarkReadStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Mark a message as read
      tags:
      - Message
  /message/presence:
    post:
      consumes:
      - application/json
      description: Set chat presence
      parameters:
      - description: Set chat presence
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_message_service.ChatPresenceStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set chat presence
      tags:
      - Message
  /message/react:
    post:
      consumes:
      - application/json
      description: React a message
      parameters:
      - description: React a message
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_message_service.ReactStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: React a message
      tags:
      - Message
  /message/status:
    post:
      consumes:
      - application/json
      description: Get message status
      parameters:
      - description: Get message status
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_message_service.MessageStatusStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get message status
      tags:
      - Message
...
swagger: "2.0"